debug

+++ dataloader.py
@@ -41,10 +41,10 @@ class FloorPlanDataset(Dataset):
 
         #plt.imsave(self.images[index],arr=gt/255)       
         if self.transform is not None:
-            image = torch.tensor([image],dtype=torch.long)
-            gt = torch.tensor([gt],dtype=torch.long)
+            image = torch.tensor([image])
+            gt = torch.tensor([gt])
         
-        
+        gt = gt.to(dtype=torch.long)
         #print(np.shape(image),np.shape(gt))
         
         return image, gt
